pyproject.toml
[project]
name = "flux-imagenet-framework"
version = "0.1.0"
description = "Generate images with Hugging Face FLUX conditioned on ImageNet class labels"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "diffusers>=0.30.0",
  "transformers>=4.44.0",
  "accelerate>=0.33.0",
  "sentencepiece>=0.2.0",
  "Pillow>=10.0.0",
  "imagenet-simple-labels>=1.0.0"
]

[project.scripts]
flux-imagenet = "flux_imagenet.cli:main"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
